Enum ritelinked::linked_hash_map::Entry

pub enum Entry<'a, K, V, S> {
    Occupied(OccupiedEntry<'a, K, V>),
    Vacant(VacantEntry<'a, K, V, S>),
}

Variants

Occupied(OccupiedEntry<'a, K, V>)

Vacant(VacantEntry<'a, K, V, S>)

Implementations

impl<'a, K, V, S> Entry<'a, K, V, S>

pub fn or_insert(self, default: V) -> &'a mut Vwhere
K: Hash,
S: BuildHasher,

If this entry is vacant, inserts a new entry with the given value and returns a reference to it.

If this entry is occupied, this method moves the occupied entry to the back of the internal linked list and returns a reference to the existing value.

pub fn or_insert_with<F: FnOnce() -> V>(self, default: F) -> &'a mut Vwhere
K: Hash,
S: BuildHasher,

Similar to Entry::or_insert, but accepts a function to construct a new value if this entry is vacant.

pub fn key(&self) -> &K

pub fn and_modify<F>(self, f: F) -> Selfwhere
F: FnOnce(&mut V),