Struct ritelinked::LinkedHashSet

pub struct LinkedHashSet<T, S = DefaultHashBuilder> { /* private fields */ }

Implementations

impl<T: Hash + Eq> LinkedHashSet<T, DefaultHashBuilder>

pub fn new() -> LinkedHashSet<T, DefaultHashBuilder>

pub fn with_capacity(capacity: usize) -> LinkedHashSet<T, DefaultHashBuilder>

impl<T, S> LinkedHashSet<T, S>

pub fn capacity(&self) -> usize

pub fn iter(&self) -> Iter<'_, T>

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn drain(&mut self) -> Drain<'_, T>

pub fn clear(&mut self)

impl<T, S> LinkedHashSet<T, S>where
T: Eq + Hash,
S: BuildHasher,

pub fn with_hasher(hasher: S) -> LinkedHashSet<T, S>

pub fn with_capacity_and_hasher(
capacity: usize,
hasher: S
) -> LinkedHashSet<T, S>

pub fn hasher(&self) -> &S

pub fn reserve(&mut self, additional: usize)

pub fn try_reserve(&mut self, additional: usize) -> Result<(), TryReserveError>

pub fn shrink_to_fit(&mut self)

pub fn difference<'a>(
&'a self,
other: &'a LinkedHashSet<T, S>
) -> Difference<'a, T, S>

pub fn symmetric_difference<'a>(
&'a self,
other: &'a LinkedHashSet<T, S>
) -> SymmetricDifference<'a, T, S>

pub fn intersection<'a>(
&'a self,
other: &'a LinkedHashSet<T, S>
) -> Intersection<'a, T, S>

pub fn union<'a>(&'a self, other: &'a LinkedHashSet<T, S>) -> Union<'a, T, S>

pub fn contains<Q: ?Sized>(&self, value: &Q) -> boolwhere
T: Borrow<Q>,
Q: Hash + Eq,

pub fn get<Q: ?Sized>(&self, value: &Q) -> Option<&T>where
T: Borrow<Q>,
Q: Hash + Eq,

pub fn get_or_insert(&mut self, value: T) -> &T

pub fn get_or_insert_with<Q: ?Sized, F>(&mut self, value: &Q, f: F) -> &Twhere
T: Borrow<Q>,
Q: Hash + Eq,
F: FnOnce(&Q) -> T,

pub fn is_disjoint(&self, other: &LinkedHashSet<T, S>) -> bool

pub fn is_subset(&self, other: &LinkedHashSet<T, S>) -> bool

pub fn is_superset(&self, other: &LinkedHashSet<T, S>) -> bool

pub fn insert(&mut self, value: T) -> bool

Inserts the given value into the set.

If the set did not have this value present, inserts it at the back of the internal linked list and returns true, otherwise it moves the existing value to the back of the internal linked list and returns false.

pub fn replace(&mut self, value: T) -> Option<T>

Adds the given value to the set, replacing the existing value.

If a previous value existed, returns the replaced value. In this case, the value’s position in the internal linked list is not changed.

pub fn remove<Q: ?Sized>(&mut self, value: &Q) -> boolwhere
T: Borrow<Q>,
Q: Hash + Eq,

pub fn retain<F>(&mut self, f: F)where
F: FnMut(&T) -> bool,

pub fn take<Q: ?Sized>(&mut self, value: &Q) -> Option<T>where
T: Borrow<Q>,
Q: Hash + Eq,

pub fn front(&self) -> Option<&T>

pub fn pop_front(&mut self) -> Option<T>

pub fn back(&self) -> Option<&T>

pub fn pop_back(&mut self) -> Option<T>

pub fn to_front<Q: ?Sized>(&mut self, value: &Q) -> boolwhere
T: Borrow<Q>,
Q: Hash + Eq,

pub fn to_back<Q: ?Sized>(&mut self, value: &Q) -> boolwhere
T: Borrow<Q>,
Q: Hash + Eq,